+++ src/test/java/by/pizzatempo/pages/products/DrinkPage.java
@@ -1,10 +1,8 @@
-package by.pizzatempo.pages;
+package by.pizzatempo.pages.products;
 
-import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
-import org.openqa.selenium.support.ui.ExpectedConditions;
 
 import java.util.List;
