+++ src/test/java/by/pizzatempo/pages/products/PizzaPage.java
@@ -1,6 +1,5 @@
-package by.pizzatempo.pages;
+package by.pizzatempo.pages.products;
 
-import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
